Use the rules from doc-linter-rules and clean up non relevant Unit Tests (#20)
* Use the rules from doc-linter-rules
