--- add-on/tests/testlist.js
@@ -1,31 +1,0 @@
-/* eslint-disable */
-const testList = [
-  "old-urls.js",
-  "absolute-urls-for-internal-links.js",
-  "different-locale-links.js",
-  "empty-elements.js",
-  "empty-brackets.js",
-  "style-attribute.js",
-  "name-attribute.js",
-  "span-count.js",
-  "pre-without-class.js",
-  "summary-heading.js",
-  "example-colon-heading.js",
-  "alert-print-in-code.js",
-  "shell-prompts.js",
-  "html-comments.js",
-  "font-elements.js",
-  "http-links.js",
-  "data-macro-note.js",
-  "unnecessary-macro-params.js",
-  "invalid-macros.js",
-  "incorrectly-wrapped-sidebar-macros.js",
-  "macro-syntax-error.js",
-  "wrong-highlighted-line.js",
-  "api-syntax-headlines.js",
-  "wrong-syntax-class.js",
-  "code-in-pre.js",
-  "line-length-in-pre.js",
-  "url-in-link-title.js",
-  "article-length.js"
-];
